moved two components
