+++ packages/components/modules/activity-log/web/ActivityLogComponent/index.tsx
@@ -8,8 +8,8 @@ import { Box, Typography } from '@mui/material'
 import { useForm } from 'react-hook-form'
 
 import { SearchNotFoundState } from '../../../__shared__/web'
-import DateFilterChip from '../../DateFilterChip'
 import { FetchParameters, useActivityLogs } from '../../common'
+import DateFilterChip from './DateFilterChip'
 import EventFilterChip from './EventFilterChip'
 import DefaultLogGroups from './LogGroups'
 import { ActivityLogComponentProps, EventFilterOption } from './types'
